backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Serwis API Flask backend dla projektu TEG obsługujący logikę biznesową i przetwarzanie danych"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "pyyaml>=6.0",
    "flask-cors>=4.0.0",
]
